requirements.txt
aerich==0.5.8
aiofiles==0.7.0
aiosqlite==0.17.0
asgiref==3.4.1
asyncpg==0.24.0
autopep8==1.5.7
click==8.0.1
databases==0.5.2
ddlparse==1.10.0
dictdiffer==0.9.0
environs==9.3.3
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
iso8601==0.1.16
marshmallow==3.13.0
multidict==5.1.0
pycodestyle==2.7.0
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dotenv==0.19.0
pytz==2021.1
sanic==21.6.2
sanic-envconfig==1.3.4
sanic-routing==0.7.1
SQLAlchemy==1.4.25
standard==1.0.3
termcolor==1.1.0
toml==0.10.2
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
ujson==4.2.0
uvicorn==0.15.0
uvloop==0.16.0
websockets==10.0
